+++ tests/specs/converters/Scss.spec.js
@@ -1,17 +1,18 @@
-const ScssConverter = require('../../../src/converters/Scss')
-const testConfig = require('../../tailwind.config')
+import ScssConverter from '../../../src/converters/Scss'
+import testConfig from '../../tailwind.config'
+import { resolveConfig } from '../../../src/converters/utils'
 
 describe('Scss converter', () => {
   it('Converts to nested map', () => {
     const converter = new ScssConverter({
-      config: testConfig
+      config: resolveConfig(testConfig)
     })
     expect(converter.convert()).toMatchSnapshot()
   })
 
   it('Converts to flat variables', () => {
     const converter = new ScssConverter({
-      config: testConfig,
+      config: resolveConfig(testConfig),
       flat: true
     })
     expect(converter.convert()).toMatchSnapshot()
@@ -19,7 +20,7 @@ describe('Scss converter', () => {
 
   it('Converts to flat variables with prefix', () => {
     const converter = new ScssConverter({
-      config: testConfig,
+      config: resolveConfig(testConfig),
       flat: true,
       prefix: 'tw'
     })
@@ -28,7 +29,7 @@ describe('Scss converter', () => {
 
   it('Converts to nested map with prefix', () => {
     const converter = new ScssConverter({
-      config: testConfig,
+      config: resolveConfig(testConfig),
       prefix: 'tw'
     })
     expect(converter.convert()).toMatchSnapshot()