+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.0.0-beta.1"></a>
+# [2.0.0-beta.1](https://github.com/dobromir-hristov/tailwindcss-export-config/compare/v2.0.0-beta.0...v2.0.0-beta.1) (2019-04-24)
+
+
+### Bug Fixes
+
+* move tailwind to a dependency ([bf6e34a](https://github.com/dobromir-hristov/tailwindcss-export-config/commit/bf6e34a))
+
+
+
+<a name="2.0.0-beta.0"></a>
+# [2.0.0-beta.0](https://github.com/dobromir-hristov/tailwindcss-export-config/compare/v1.0.4...v2.0.0-beta.0) (2019-04-24)
+
+
+### Features
+
+* update library to work with Tailwind 1.x ([e005737](https://github.com/dobromir-hristov/tailwindcss-export-config/commit/e005737))
+
+
+### BREAKING CHANGES
+
+* Works with new tailwind config structure
+
+
+
 <a name="1.0.5"></a>
 ## [1.0.5](https://github.com/dobromir-hristov/tailwindcss-export-config/compare/v1.0.4...v1.0.5) (2019-04-24)